fix: Resolve subproject tax display and replace all Float with Decimal
- Remove tax display from subproject cost summaries (show subtotal only)
- Update CostSummary component to conditionally show tax vs subtotal
- Replace all Float fields in models.py with proper Decimal precision
- Clean up subproject cost-summary API to remove unused tax calculations
- Fix naming conflicts in SubprojectDetail component

Database schema changes:
- Currency fields: Numeric(10,2) for up to $99,999,999.99
- Tax rates: Numeric(5,4) for up to 99.9999% precision
- Hours/quantities: Numeric(8,2) for up to 999,999.99 units
- Maintains financial precision, eliminates floating point errors

+++ backend/app/models/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Float, DateTime, Text, ForeignKey, Boolean, Numeric
+from sqlalchemy import Column, Integer, String, DateTime, Text, ForeignKey, Boolean, Numeric
 from sqlalchemy.orm import relationship
 from sqlalchemy.sql import func
 from app.db.database import Base
@@ -70,12 +70,12 @@ class Project(Base):
     status = Column(String, default=PROJECT_STATUS_PLANNING)  # planning, in_progress, completed, on_hold
     start_date = Column(DateTime)
     end_date = Column(DateTime)
-    estimated_cost = Column(Float, default=0.0)
-    actual_cost = Column(Float, default=0.0)
-    labor_cost = Column(Float, default=0.0)
-    material_cost = Column(Float, default=0.0)
-    permit_cost = Column(Float, default=0.0)
-    other_cost = Column(Float, default=0.0)
+    estimated_cost = Column(Numeric(10,2), default=0.0)
+    actual_cost = Column(Numeric(10,2), default=0.0)
+    labor_cost = Column(Numeric(10,2), default=0.0)
+    material_cost = Column(Numeric(10,2), default=0.0)
+    permit_cost = Column(Numeric(10,2), default=0.0)
+    other_cost = Column(Numeric(10,2), default=0.0)
     sales_tax_rate = Column(Numeric(5, 4), default=0.0)
     sales_tax_amount = Column(Numeric(10, 2), default=0.0)
     is_tax_exempt = Column(Boolean, default=False)
@@ -102,8 +102,8 @@ class Task(Base):
     description = Column(Text)
     status = Column(String, default=TASK_STATUS_PENDING)  # pending, in_progress, completed
     priority = Column(String, default=TASK_PRIORITY_MEDIUM)  # low, medium, high
-    estimated_hours = Column(Float, default=0.0)
-    actual_hours = Column(Float, default=0.0)
+    estimated_hours = Column(Numeric(8,2), default=0.0)
+    actual_hours = Column(Numeric(8,2), default=0.0)
     due_date = Column(DateTime)
     completed_at = Column(DateTime)
     created_at = Column(DateTime(timezone=True), server_default=func.now())
@@ -122,10 +122,10 @@ class Invoice(Base):
     invoice_number = Column(String, unique=True, index=True)
     title = Column(String)
     description = Column(Text)
-    amount = Column(Float)
-    tax_rate = Column(Float, default=0.0)
-    tax_amount = Column(Float, default=0.0)
-    total_amount = Column(Float)
+    amount = Column(Numeric(10,2))
+    tax_rate = Column(Numeric(5,4), default=0.0)
+    tax_amount = Column(Numeric(10,2), default=0.0)
+    total_amount = Column(Numeric(10,2))
     status = Column(String, default=INVOICE_STATUS_DRAFT)  # draft, sent, paid, overdue
     issue_date = Column(DateTime)
     due_date = Column(DateTime)
@@ -177,7 +177,7 @@ class MaterialEntry(Base):
     description = Column(String, index=True)
     unit = Column(String)
     category = Column(String)
-    unit_price = Column(Float)
+    unit_price = Column(Numeric(10,2))
     created_at = Column(DateTime(timezone=True), server_default=func.now())
     
     # Foreign key
@@ -193,8 +193,8 @@ class MaterialItem(Base):
     id = Column(Integer, primary_key=True, index=True)
     description = Column(String)
     unit = Column(String)
-    quantity = Column(Float)
-    unit_cost = Column(Float)
+    quantity = Column(Numeric(8,2))
+    unit_cost = Column(Numeric(10,2))
     category = Column(String)
     created_at = Column(DateTime(timezone=True), server_default=func.now())
     updated_at = Column(DateTime(timezone=True), onupdate=func.now())
@@ -212,8 +212,8 @@ class LaborItem(Base):
     id = Column(Integer, primary_key=True, index=True)
     role = Column(String)
     number_of_workers = Column(Integer)
-    hourly_rate = Column(Float)
-    hours = Column(Float)
+    hourly_rate = Column(Numeric(8,2))
+    hours = Column(Numeric(8,2))
     created_at = Column(DateTime(timezone=True), server_default=func.now())
     updated_at = Column(DateTime(timezone=True), onupdate=func.now())
     
@@ -229,7 +229,7 @@ class PermitItem(Base):
 
     id = Column(Integer, primary_key=True, index=True)
     description = Column(String)
-    cost = Column(Float)
+    cost = Column(Numeric(10,2))
     issued_date = Column(DateTime)
     expiration_date = Column(DateTime)
     notes = Column(Text)
@@ -248,7 +248,7 @@ class OtherCostItem(Base):
 
     id = Column(Integer, primary_key=True, index=True)
     description = Column(String)
-    cost = Column(Float)
+    cost = Column(Numeric(10,2))
     notes = Column(Text)
     created_at = Column(DateTime(timezone=True), server_default=func.now())
     updated_at = Column(DateTime(timezone=True), onupdate=func.now())
